requirements.txt
Flask==3.0.0
gunicorn
python-dotenv
waitress
azure-ai-inference
azure-core
httpx
selenium
pandas
langchain
BeautifulSoup4
mcp
openpyxl
pytest
seaborn
matplotlib
scikit-learn